osd/ReplicatedPG: basic flush and evict agent functionality
This is very basic flush and evict functionality for the tiering agent.

The flush policy is very simple: if we are above the threshold and the
object is dirty, and not super young, flush it.  This is not too braindead
of a policy (although we could clearly do something smarter).

The evict policy is pretty braindead: evict the object if it is clean and
we are over our full threshold.  This needs to be augmented to estimate
age based on the HitSets.

Signed-off-by: Sage Weil <sage@inktank.com>